server/service_asc_aed/model/asc_aed/requirement.txt
pillow==11.0.0
natsort==7.1.1
werkzeug==3.1.3
librosa==0.10.1
torch==2.3.1 
torchvision==0.18.1 
torchaudio==2.3.1 
torchlibrosa==0.1.0
pandas==2.2.3
h5py==3.12.1 
